requirements.txt
flask
flask_sqlalchemy
python-dotenv
geoip2
